backend/app/services/storage.py
from typing import Dict, Any, List
import boto3
from google.cloud import storage
from app.config import settings

class StorageService:
    """
    File storage helpers for S3/GCS
    """
    
    def __init__(self):
        self.s3_client = None
        self.gcs_client = None
    
    async def store_file(self, file_path: str, content: bytes) -> str:
        """Store file in cloud storage"""
        # TODO: Implement file storage logic
        return f"stored://{file_path}"
    
    async def retrieve_file(self, file_path: str) -> bytes:
        """Retrieve file from cloud storage"""
        # TODO: Implement file retrieval
        return b""
    
    async def delete_file(self, file_path: str) -> bool:
        """Delete file from cloud storage"""
        # TODO: Implement file deletion
        return True
    
    async def list_files(self, prefix: str = "") -> List[str]:
        """List files in storage"""
        # TODO: Implement file listing
        return []
